requirements.txt
Flask
Werkzeug
gunicorn
Pillow
captcha
